Gradio ui clean and refactor
